+++ integration-tests/http/__tests__/cart/store/cart.spec.ts
@@ -832,8 +832,10 @@ medusaIntegrationTestRunner({
                   tax_lines: [
                     {
                       code: "CADEFAULT",
+                      data: null,
                       description: "CA Default Rate",
                       id: expect.any(String),
+                      metadata: null,
                       provider_id: "system",
                       rate: 5,
                     },
@@ -954,8 +956,10 @@ medusaIntegrationTestRunner({
                   tax_lines: [
                     {
                       code: "CADEFAULT",
+                      data: null,
                       description: "CA Default Rate",
                       id: expect.any(String),
+                      metadata: null,
                       provider_id: "system",
                       rate: 5,
                     },
@@ -4227,11 +4231,10 @@ medusaIntegrationTestRunner({
           ).data.product
 
           let updated = await api.post(
-            `/store/carts/${cart.id}/line-items?fields=+items.is_giftcard`,
+            `/store/carts/${cart.id}/line-items`,
             { variant_id: giftCardProduct.variants[0].id, quantity: 1 },
             storeHeaders
           )
-
           expect(updated.status).toEqual(200)
           expect(updated.data.cart).toEqual(
             expect.objectContaining({