+++ packages/admin/dashboard/src/routes/product-types/product-type-list/product-type-list.tsx
@@ -1,18 +1,20 @@
-import { SingleColumnPage } from "../../../components/layout/pages"
-import { useExtension } from "../../../providers/extension-provider"
+import { CORE_LAYOUT_IDS } from "@medusajs/admin-shared"
+
+import { LayoutComposer } from "../../../components/layout-composer"
 import { ProductTypeListTable } from "./components/product-type-list-table"
 
 export const ProductTypeList = () => {
-  const { getWidgets } = useExtension()
-
   return (
-    <SingleColumnPage
-      widgets={{
-        after: getWidgets("product_type.list.after"),
-        before: getWidgets("product_type.list.before"),
+    <LayoutComposer
+      widgetsZonePrefix="product_type.list"
+      preferredLayoutId={CORE_LAYOUT_IDS.SINGLE_COLUMN}
+      sections={{
+        main: (
+          <LayoutComposer.Entry id="ProductTypeListTable">
+            <ProductTypeListTable />
+          </LayoutComposer.Entry>
+        ),
       }}
-    >
-      <ProductTypeListTable />
-    </SingleColumnPage>
+    />
   )
 }