+++ packages/admin/dashboard/src/routes/shipping-profiles/shipping-profiles-list/shipping-profile-list.tsx
@@ -1,18 +1,20 @@
-import { SingleColumnPage } from "../../../components/layout/pages"
-import { useExtension } from "../../../providers/extension-provider"
+import { CORE_LAYOUT_IDS } from "@medusajs/admin-shared"
+
+import { LayoutComposer } from "../../../components/layout-composer"
 import { ShippingProfileListTable } from "./components/shipping-profile-list-table"
 
 export const ShippingProfileList = () => {
-  const { getWidgets } = useExtension()
-
   return (
-    <SingleColumnPage
-      widgets={{
-        before: getWidgets("shipping_profile.list.before"),
-        after: getWidgets("shipping_profile.list.after"),
+    <LayoutComposer
+      widgetsZonePrefix="shipping_profile.list"
+      preferredLayoutId={CORE_LAYOUT_IDS.SINGLE_COLUMN}
+      sections={{
+        main: (
+          <LayoutComposer.Entry id="ShippingProfileListTable">
+            <ShippingProfileListTable />
+          </LayoutComposer.Entry>
+        ),
       }}
-    >
-      <ShippingProfileListTable />
-    </SingleColumnPage>
+    />
   )
 }