+++ packages/modules/providers/payment-stripe/src/types/index.ts
@@ -1,3 +1,5 @@
+import Stripe from "stripe"
+
 export interface StripeOptions {
   /**
    * The API key for the Stripe account
@@ -23,6 +25,11 @@ export interface StripeOptions {
    * Set the number of days before an OXXO payment expires
    */
   oxxoExpiresDays?: number
+  /**
+   * List of asynchronous payment method types. If not provided, all payment method types will be considered synchronous.
+   * Asynchronous payment methods are allowed to generate an order when the Stripe status is `pending`.
+   */
+  asyncPaymentMethodTypes?: Stripe.PaymentMethod.Type[]
 }
 
 export interface PaymentIntentOptions {